+++ source/modules/database/BaseDatabase.py
@@ -1,14 +1,13 @@
-# strategies/base_actions.py
+# IntelliTrader\source\modules\database\BaseDatabase.py
+
 from abc import ABC, abstractmethod
+from source.modules.shared.shared_handler import SharedHandler
 from sys import modules
 
-class BaseDatabase(ABC):
+class BaseDatabase(ABC, SharedHandler):
     def __init__(self, connection, modules):
         self.connection = connection
         self.modules = modules
-        self.table_name = None
-        self.region_name = None
-        self.dynamodb_client = None # DynamoDBClient(self.table_name, self.region_name)
 
     @abstractmethod
     def initialize(self):