+++ source/modules/database/_database_create_table.py
@@ -3,74 +3,112 @@
 from source.constants.constants import *
 from source.enumerations.enums import *
 from source.utils.logging_utils import *
-from source.aws.DynamoDB.aws_dynamodb import AWSDynamoDB
-from botocore.client import ClientError
+from source.modules.database.comparison_operator import ComparisonOperators
+
+from boto3.exceptions import botocore
+from botocore.exceptions import ClientError
+from boto3.dynamodb.conditions import Key, Attr
+import boto3
 
 class DatabaseCreateTable:
-    def __init__(self, configuration):
-        self.configuration = configuration
+    def __init__(self, table_configuration, table_name):
 
+        self.table_configuration = table_configuration
+        self.table_name          = table_name
+        self.dynamodb_resource   = boto3.resource('dynamodb', region_name=REGION_NAME)
+        self.dynamodb_table      = self.dynamodb_resource.Table(self.table_name)
+
+        
     def initialize(self):
         return self.create_tables()
-    
+
+
     def create_tables(self):
-        # Create Required Tables
-        tables_list = {
-            'table_1': 'Topics'
-        }
 
-        table_properties = {}
-        for table_name, display_name in tables_list.items():
-            table_properties[table_name] = self.build_table_definition(display_name)  
+        """ Check If Table Exists """
+        table_exists = self.check_table_exists()  
 
-        object_dynamodb_create_handler = AWSDynamoDB()
-        for table_name, properties in table_properties.items():
+        if not table_exists:
+
+            table_properties = {}
+            table_properties = self.build_table_definition()
+            
             try:
-                object_dynamodb_create_handler.create_table(properties)
-            except ClientError as e:
-                log_error(f"Error creating table '{table_name}': {e}")
+                log_info(f"Creating table '{self.table_name}'...")
+                response                  = self.dynamodb_resource.create_table(
+                    TableName             = self.table_name,
+                    KeySchema             = table_properties['table_schema'],
+                    AttributeDefinitions  = table_properties['attribute_definitions'],
+                    ProvisionedThroughput = {
+                    'ReadCapacityUnits': 5,
+                    'WriteCapacityUnits': 5
+                    }
+                )
+                response.wait_until_exists()
+                log_info(f"Create '{self.table_name}' Table ...COMPLETE!")
                 
+            except ClientError as e:
+                log_error(f"Error creating table {self.table_name}. Here's why: {e.response["Error"]["Code"]}: {e.response["Error"]["Message"]}")
+                raise
 
-    def build_table_definition(self, table_name):
-        """
-        Defines the table structure for either Topics or Alerts table.
+        else:
+            log_info(f"Table '{self.table_name}' already exists.")
+                                
 
-        Args:
-            table_name (str): The name of the table to be created ("Topics" or "Alerts").
+    def build_table_definition(self):
+        
+        table_config = self.table_configuration
+        table_key_name = table_config['table_key']
+        hash_key_name  = table_config['hash_key']
+        sort_key_name  = table_config['sort_key'] 
 
-        Returns:
-            dict: A dictionary containing the table properties for creating the table.
-        """
-
-        table_properties = {
-            'table_name': table_name
+        table_definition = {
+            'table_name': table_key_name,
+            'table_schema': [
+                {
+                    'AttributeName': hash_key_name,
+                    'KeyType': 'HASH'
+                },
+                # Only include sort key definition if sort_key_name is not empty
+                *([
+                    {
+                        'AttributeName': sort_key_name,
+                        'KeyType': 'RANGE'
+                    }
+                ] if sort_key_name else [])
+            ],
+            'attribute_definitions': [
+                {
+                    'AttributeName': hash_key_name,
+                    'AttributeType': 'S'
+                },
+                # Only include sort key attribute definition if sort_key_name is not empty
+                *([
+                    {
+                        'AttributeName': sort_key_name,
+                        'AttributeType': 'S'
+                    }
+                ] if sort_key_name else [])
+            ]
         }
 
-        ######################################
-        # Define table properties for topics #
-        ######################################
+        return table_definition
 
-        if table_name == "Topics":
 
-            table_properties['table_schema'] = [
-                {
-                    'AttributeName': 'Id',
-                    'KeyType': 'HASH'  
-                },
-                {
-                    'AttributeName': 'TopicName',
-                    'KeyType': 'RANGE' 
-                }
-            ]
-            table_properties['attribute_definitions'] = [
-                {
-                    'AttributeName': 'Id',
-                    'AttributeType': 'S'  
-                },
-                {
-                    'AttributeName': 'TopicName',
-                    'AttributeType': 'S'  
-                }
-            ]
 
-        return table_properties
+    def check_table_exists(self):
+        try:
+            self.dynamodb_table.load()
+            exists = True
+        except ClientError as e:
+            if e.response['Error']['Code'] == 'ResourceNotFoundException':
+                exists = False
+            else:
+                log_error(
+                    "Couldn't check for existence of %s. Here's why: %s: %s",
+                    self.table_name,
+                    e.response["Error"]["Code"],
+                    e.response["Error"]["Message"],
+                )
+                raise
+        return exists