+++ source/modules/database/comparison_operator.py
@@ -1,0 +1,10 @@
+
+class ComparisonOperators:
+    EQ = 'eq'  # Equal to
+    LT = 'lt'  # Less than
+    GT = 'gt'  # Greater than
+    GTE = 'gte'  # Greater than or equal to
+    LTE = 'lte'  # Less than or equal to
+    NE = 'ne'  # Not equal to
+    BETWEEN = 'between'  # Between
+