+++ source/modules/strategy/BaseStrategy.py
@@ -1,7 +1,8 @@
-# strategies/BaseStrategy.py
+# IntelliTrader\source\modules\strategy\BaseStrategy.py
+
 from abc import ABC, abstractmethod
-from sys import modules
 from source.modules.shared.shared_handler import SharedHandler
+from sys import modules
 
 class BaseStrategy(ABC, SharedHandler):
     def __init__(self, connection, modules):