+++ source/modules/strategy/_strategy_candlesticks.py
@@ -2,7 +2,6 @@
 
 import asyncio
 import datetime
-import shutil
 import time
 import os
 
@@ -24,6 +23,7 @@ class StrategyCandlesticks:
         self.timeframe                    = parameters.get('timeframe')
         self.historical_data_subscription = parameters.get('historical_data_subscription')
         self.cache_path                   = CACHE_PATH
+        self.market_start_time            = datetime.time(hour=9, minute=15)
         
     def initialize(self):
         return self.scan_watchlist_stocks()
@@ -54,8 +54,14 @@ class StrategyCandlesticks:
 
         # Get current date
         now = datetime.datetime.now()
-        formatted_date = now.strftime("%Y-%m-%d")
+        previous_day = now - datetime.timedelta(days=1)        
+        current_time = now.time()
 
+        if current_time <= self.market_start_time:
+            formatted_date = previous_day.strftime("%Y-%m-%d")
+        else:
+            formatted_date = now.strftime("%Y-%m-%d")
+            
         # Check if cache directory needs to be updated
         if not os.path.exists(self.cache_path) or self.cache_path.split("/")[-1] != formatted_date:
             # Cache path doesn't exist or date has changed, create new directory
@@ -65,20 +71,12 @@ class StrategyCandlesticks:
             except OSError as e:
                 print(f"Error creating directory: {e}")
 
-            # Remove old cache directories
-            old_cache_dirs = [os.path.join(CACHE_PATH, dir) for dir in os.listdir(CACHE_PATH) if dir != formatted_date]
-            for dir in old_cache_dirs:
-                try:
-                    if os.path.isdir(dir):
-                        shutil.rmtree(dir)  
-                except OSError as e:
-                    print(f"Error removing directory {dir}: {e}")
-
         # Individual cache objects for timeframes
         timeframe_caches = {}
 
         # Timeframes to cache
-        timeframes_to_cache = [DAY, WEEK, MONTH, TODAY_5M, TODAY_15M, TODAY_30M, TODAY_60M]
+        timeframes_to_cache = [DAY, WEEK, MONTH, TODAY_5M] 
+        """ TODAY_15M, TODAY_30M, TODAY_60M """
       
         for timeframe in timeframes_to_cache:
             # Create separate cache object for each timeframe
@@ -93,8 +91,7 @@ class StrategyCandlesticks:
             if timeframe.startswith("today"):
 
                 minutes = int((timeframe.split("today")[1]).split("minute")[0])
-                current_time = now.time()
-                target_time = datetime.datetime.combine(now.date(), datetime.time(hour=9, minute=15))  
+                target_time = datetime.datetime.combine(now.date(), self.market_start_time)  
                 target_time += datetime.timedelta(minutes=minutes)  
 
                 # Check if current time is past the target time
@@ -107,10 +104,18 @@ class StrategyCandlesticks:
                         candlestick_data[timeframe] = candles
                         timeframe_caches[timeframe].set(cache_key, candles)  
                 else:
-                    print(f"Waiting for {timeframe} data to be available...")
-                    
+                    cached_data = timeframe_caches[timeframe].get(cache_key)
+                    if cached_data is not None:
+                        candlestick_data[timeframe] = cached_data
+                    elif current_time <= self.market_start_time:
+                        candles = await self.fetch_ohlc_async(self.trading_exchange, self.trading_symbol, self.trading_token, timeframe)
+                        candlestick_data[timeframe] = candles
+                        timeframe_caches[timeframe].set(cache_key, candles)   
+                    else:    
+                        log_warn(f"Waiting for {timeframe} live data to be available after market opens, using previous day data...")
+                        
             else:
-                # TODAY, WEEKLY and MONTHLY timeframe, proceed with regular cache operations
+                # DAILY, WEEKLY and MONTHLY timeframe, proceed with regular cache operations
                 cached_data = timeframe_caches[timeframe].get(cache_key)
                 if cached_data is not None:
                     candlestick_data[timeframe] = cached_data